keep PIT entry until all upstreams timeout
This helps strategy to determine when an Interest is sent.

A (possibly bad) side effect:
1. an Interest with AnswerOriginKind=0 (bypass caches) is satisfied by upstream A but still pending on upstream B
2. a similar Interest comes in
3. previously, a new PIT entry is created so that strategy can forward to both A and B; now, strategy can only forward to A
note: strategy->PropagateNewInterest is still invoked now due to `ie->native()->strategy.renewals = 0;`

Remaining issue: when an Interest is satisfied by a CO from InternalClient, Strategy::WillSatisfyPendingInterest won't be invoked so that that Interest would be considered unanswered and invoke Strategy::WillEraseTimedOutPendingInterest.